feat: 딥링크 처리 로직 이전 및 앱 종료 방식 개선
딥링크를 통한 화면 이동 로직을 최상위 `FestabookScreen`에서 `MainScreen`으로 이전하고, `Activity` 종료 방식을 콜백 기반으로 구조화하여 제어 흐름을 개선했습니다.

- **딥링크 처리 로직 이전**:
    - `FestabookScreen`에 위치하던 `RememberDeepLinkHandler`와 관련 내비게이션 처리 함수(`handleNavigation`)를 `MainScreen`으로 이동했습니다.
    - 이를 통해 메인 화면 진입 시 공지사항 확장 및 뉴스 탭 이동 등의 딥링크 이벤트를 보다 적절한 생명주기 내에서 처리하도록 변경했습니다.

- **앱 종료 로직 개선**:
    - `MainActivity`에서 `finish()`를 호출하는 `onAppFinish` 콜백을 `FestabookScreen` 및 하위 내비게이션 그래프(`mainNavGraph`, `splashNavGraph`)로 전달하도록 수정했습니다.
    - 기존에 `festabookNavigator.popBackStack()`을 이용해 간접적으로 앱을 종료하던 방식을 명시적인 콜백 호출 방식으로 변경하여 안정성을 높였습니다.

- **기타 수정 사항**:
    - `MainActivity.newIntent` 호출 시 사용되는 플래그를 `FLAG_ACTIVITY_NEW_TASK` 및 `FLAG_ACTIVITY_CLEAR_TASK`로 변경하여 인텐트 발생 시 액티비티 스택을 초기화하도록 수정했습니다.
    - `FestabookScreen`에서 더 이상 직접 사용하지 않는 `NewsViewModel`과 `MainViewModel` 주입 로직을 제거했습니다.
